refactor(core/domain): split feedback misses subpackage
Separate taxonomy, JSONL store, and benchmark export into focused modules
under feedback/misses/ while keeping the public import paths unchanged.
